Pull request release 1.1.0

+++ squalaetp/management/commands/importcorvet.py
@@ -6,7 +6,7 @@ from ._excel_squalaetp import ExcelSqualaetp
 from squalaetp.models import Xelon, Sivin
 from psa.models import Corvet
 from utils.scraping import ScrapingCorvet, ScrapingSivin
-from utils.django.validators import xml_parser, xml_sivin_parser
+from utils.django.validators import xml_parser, xml_sivin_parser, VIN_PSA_REGEX
 from utils.django.models import defaults_dict
 from utils.conf import XLS_SQUALAETP_FILE
 
@@ -51,61 +51,62 @@ class Command(BaseCommand):
             squalaetp = ExcelSqualaetp(XLS_SQUALAETP_FILE)
             xelon_list = squalaetp.xelon_number_list()
             xelons = Xelon.objects.filter(
-                numero_de_dossier__in=xelon_list, vin__regex=r'^V((F[37])|(R[137]))\w{14}$',
+                numero_de_dossier__in=xelon_list, vin__regex=VIN_PSA_REGEX,
                 vin_error=False, corvet__isnull=True)
             self.stdout.write(f"[IMPORT_CORVET] Xelon number = {xelons.count()}")
             nb_file = self._import_corvet(xelons)
             self.stdout.write(self.style.SUCCESS(f"[IMPORT_CORVET] Import completed: NB_CORVET = {nb_file}"))
         elif options['all']:
             self.stdout.write("[IMPORT_CORVET] Waiting...")
-            corvets = Corvet.objects.filter(prods__update=True)[:200]
+            corvets = Corvet.objects.filter(opts__update=True)[:200]
             nb_file = self._import_corvet(corvets, squalaetp=False, limit=True)
             self.stdout.write(self.style.SUCCESS(f"[IMPORT_CORVET] Import completed: NB_CORVET = {nb_file}"))
         else:
             self.stdout.write("[IMPORT_CORVET] Waiting...")
-            xelons = Xelon.objects.filter(vin__regex=r'^V((F[37])|(R[137]))\w{14}$', vin_error=False).order_by('-id')
-            xelons = xelons.filter(Q(corvet__isnull=True) | Q(corvet__prods__update=True))[:200]
+            xelons = Xelon.objects.filter(vin__regex=VIN_PSA_REGEX, vin_error=False).order_by('-id')
+            xelons = xelons.filter(Q(corvet__isnull=True) | Q(corvet__opts__update=True))[:200]
             nb_file = self._import_corvet(xelons, limit=True)
             self.stdout.write(self.style.SUCCESS(f"[IMPORT_CORVET] Import completed: NB_CORVET = {nb_file}"))
 
     def _import_corvet(self, queryset, squalaetp=True, limit=False):
         nb_import = 0
-        scrap = ScrapingCorvet()
-        for query in queryset:
-            start_msg = f"{query.vin}"
-            if squalaetp:
-                start_msg = f"{query.numero_de_dossier} - {start_msg}"
-            start_time = time.time()
-            for attempt in range(2):
-                row = xml_parser(scrap.result(query.vin))
-                if scrap.ERROR or "ERREUR COMMUNICATION SYSTEME CORVET" in row:
-                    nb_import += 1
-                    delay_time = time.time() - start_time
-                    self.stdout.write(
-                        self.style.ERROR(f"{start_msg} error CORVET in {delay_time}"))
+        if queryset:
+            scrap = ScrapingCorvet()
+            for query in queryset:
+                start_msg = f"{query.vin}"
+                if squalaetp:
+                    start_msg = f"{query.numero_de_dossier} - {start_msg}"
+                start_time = time.time()
+                for attempt in range(2):
+                    row = xml_parser(scrap.result(query.vin))
+                    if scrap.ERROR or "ERREUR COMMUNICATION SYSTEME CORVET" in row:
+                        nb_import += 1
+                        delay_time = time.time() - start_time
+                        self.stdout.write(
+                            self.style.ERROR(f"{start_msg} error CORVET in {delay_time}"))
+                        break
+                    elif row and row.get('donnee_date_entree_montage'):
+                        defaults = defaults_dict(Corvet, row, "vin")
+                        obj, created = Corvet.objects.update_or_create(vin=row["vin"], defaults=defaults)
+                        obj.opts.update = False
+                        obj.opts.save()
+                        nb_import += 1
+                        delay_time = time.time() - start_time
+                        self.stdout.write(
+                            self.style.SUCCESS(f"{start_msg} updated in {delay_time}"))
+                        break
+                    if attempt:
+                        if squalaetp:
+                            query.vin_error = True
+                        Corvet.objects.filter(vin=query.vin).delete()
+                        delay_time = time.time() - start_time
+                        self.stdout.write(
+                            self.style.ERROR(f"{start_msg} error VIN in {delay_time}"))
+                if squalaetp:
+                    query.save()
+                if limit and nb_import >= 200:
                     break
-                elif row and row.get('donnee_date_entree_montage'):
-                    defaults = defaults_dict(Corvet, row, "vin")
-                    obj, created = Corvet.objects.update_or_create(vin=row["vin"], defaults=defaults)
-                    obj.prods.update = False
-                    obj.prods.save()
-                    nb_import += 1
-                    delay_time = time.time() - start_time
-                    self.stdout.write(
-                        self.style.SUCCESS(f"{start_msg} updated in {delay_time}"))
-                    break
-                if attempt:
-                    if squalaetp:
-                        query.vin_error = True
-                    Corvet.objects.filter(vin=query.vin).delete()
-                    delay_time = time.time() - start_time
-                    self.stdout.write(
-                        self.style.ERROR(f"{start_msg} error VIN in {delay_time}"))
-            if squalaetp:
-                query.save()
-            if limit and nb_import >= 200:
-                break
-        scrap.close()
+            scrap.close()
         return nb_import
 
     def _import_sivin(self, immat):
